requirements.txt
alt-profanity-check==1.8.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.6
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.1
comm==0.2.2
contourpy==1.3.1
cryptography==46.0.5
curl_cffi==0.11.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.56.0
frozendict==2.4.6
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
googleapis-common-protos==1.73.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.2.1
multidict==6.7.1
multitasking==0.0.11
narwhals==2.18.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.3
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.6.0
postgrest==2.28.2
prompt_toolkit==3.0.50
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.2.1
pyrate-limiter==3.7.0
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.1
pyzmq==26.3.0
realtime==2.28.2
referencing==0.36.2
regex==2026.2.28
requests==2.32.3
rich==14.3.3
rpds-py==0.25.1
scikit-learn==1.8.0
scipy==1.15.2
seaborn==0.13.2
sec-edgar-downloader==5.0.3
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
soupsieve==2.7
stack-data==0.6.3
storage3==2.28.2
streamlit==1.55.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.2
supabase-auth==2.28.2
supabase-functions==2.28.2
tenacity==9.1.4
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
yarl==1.23.0
yfinance==0.2.61
zstandard==0.25.0
